requirements.txt
# Core
langchain>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0

# Document loading
pypdf
unstructured
markdown

# Embeddings
sentence-transformers
torch

# Vector stores
faiss-cpu
chromadb

# Local LLM
ollama

# UI
streamlit

# Evaluation (optional but recommended)
ragas
datasets

# Utils
python-dotenv
tqdm
